Split standalone types out of BLEManager.swift, and drop the test workaround
BLEManager.swift sits at SwiftLint's 2600-line file_length ceiling, so any
addition to it fails CI. CameraMode, GoProSetting and ResponseType were declared
alongside BLEManager but are not part of it, so they move to their own file.
That brings BLEManager.swift to 2453 lines and leaves room for the rest of this
stack.

scripts/run-tests.sh existed because xcodebuild reported zero eligible
destinations when the iOS platform matching the Xcode SDK was not installed.
That was a local environment problem, now resolved, and a second test path that
duplicates target configuration would only drift. Removed in favour of the
documented invocation.

The documented invocation does have a real problem, though: with no OS= in the
destination, xcodebuild defaults to the newest installed runtime, so the
hardcoded "iPhone 16" fails on Xcode 26 where no such device exists for iOS 26.
README and CONTRIBUTING now say how to pick a destination that exists.
